+++ src/client/cypress/e2e/boreholeList.cy.js
@@ -23,38 +23,38 @@ describe("Borehole list tests", () => {
     cy.wait("@borehole");
 
     // contains names
-    cy.contains("td", "Aaliyah Swaniawski");
-    cy.contains("td", "Aaron Christiansen");
-    cy.contains("td", "Abagail Keebler");
-    cy.contains("td", "Abbey Schamberger");
+    cy.contains("td", "Aaron Beier");
+    cy.contains("td", "Aaron Johnson");
+    cy.contains("td", "Aaron Wilkinson");
+    cy.contains("td", "Abagail Hagenes");
 
     // contains date (restriction date)
-    cy.contains("td", "09.03.2022");
+    cy.contains("td", "04.02.2021");
 
     // contains total depth
-    cy.contains("td", "566.0580250276522");
-    cy.contains("td", "1'356.8295898646254");
-    cy.contains("td", "500.9678483479506");
+    cy.contains("td", "1'223.1238797414694");
+    cy.contains("td", "1'130.7863738065523");
+    cy.contains("td", "1'292.2678549272323");
 
     // sort by name descending
     cy.contains("div", "Original name").click();
     cy.wait("@borehole");
-    cy.get("tbody").children().eq(0).contains("td", "Zula Langosh");
-    cy.get("tbody").children().eq(1).contains("td", "Zora Cormier");
-    cy.get("tbody").children().eq(2).contains("td", "Zoila Toy");
+    cy.get("tbody").children().eq(0).contains("td", "Zula Will");
+    cy.get("tbody").children().eq(1).contains("td", "Zora Quigley");
+    cy.get("tbody").children().eq(2).contains("td", "Zora Mills");
 
     // sort by total depth
     cy.contains("div", "Total depth MD [m]").click();
     cy.wait("@borehole");
-    cy.get("tbody").children().eq(0).contains("td", "5.198657515085609");
-    cy.get("tbody").children().eq(1).contains("td", "5.2529545525335495");
-    cy.get("tbody").children().eq(2).contains("td", "5.30725158998149");
+    cy.get("tbody").children().eq(0).contains("td", "0.501133501763052");
+    cy.get("tbody").children().eq(1).contains("td", "1.0761823510174557");
+    cy.get("tbody").children().eq(2).contains("td", "1.5021385631999646");
 
     cy.contains("div", "Total depth MD [m]").click();
     cy.wait("@borehole");
-    cy.get("tbody").children().eq(0).contains("td", "1'999.7566677628815");
-    cy.get("tbody").children().eq(1).contains("td", "1'999.7023707254336");
-    cy.get("tbody").children().eq(2).contains("td", "1'999.6480736879855");
+    cy.get("tbody").children().eq(0).contains("td", "1'999.7876528649533");
+    cy.get("tbody").children().eq(1).contains("td", "1'999.3616966527707");
+    cy.get("tbody").children().eq(2).contains("td", "1'999.212604015699");
 
     // sort by drilling date
     cy.contains("div", "End of drilling date").click();
@@ -72,9 +72,9 @@ describe("Borehole list tests", () => {
     // sort by borehole type (column of original name)
     cy.contains("div", "Borehole type").click();
     cy.wait("@borehole");
-    cy.get("tbody").children().eq(0).contains("td", "Quincy O'Connell");
-    cy.get("tbody").children().eq(1).contains("td", "Louie Kirlin");
-    cy.get("tbody").children().eq(2).contains("td", "Ubaldo Fadel");
+    cy.get("tbody").children().eq(0).contains("td", "Brisa Jenkins");
+    cy.get("tbody").children().eq(1).contains("td", "Dillan Schmitt");
+    cy.get("tbody").children().eq(2).contains("td", "Laurine Gibson");
   });
 
   it("Boreholes are displayed in correct order with admin login", () => {
@@ -100,7 +100,7 @@ describe("Borehole list tests", () => {
         }
       });
 
-    cy.get("tbody").children().eq(0).contains("td", "31.12.2021");
+    cy.get("tbody").children().eq(0).contains("td", "01.01.2022");
     cy.get("tbody").children().eq(1).contains("td", "31.12.2021");
     cy.get("tbody").children().eq(2).contains("td", "31.12.2021");
 
@@ -145,9 +145,9 @@ describe("Borehole list tests", () => {
         }
       });
 
-    cy.get("tbody").children().eq(0).contains("td", "Aaliyah Swaniawski");
-    cy.get("tbody").children().eq(1).contains("td", "Aaron Christiansen");
-    cy.get("tbody").children().eq(2).contains("td", "Abagail Keebler");
+    cy.get("tbody").children().eq(0).contains("td", "Aaron Beier");
+    cy.get("tbody").children().eq(1).contains("td", "Aaron Johnson");
+    cy.get("tbody").children().eq(2).contains("td", "Aaron Wilkinson");
 
     // sort by borehole type
     cy.contains("th", "Borehole type").click();
@@ -166,8 +166,8 @@ describe("Borehole list tests", () => {
     // sort by total depth
     cy.contains("th", "Total depth").click();
     cy.wait("@editorBorehole");
-    cy.get("tbody").children().eq(0).contains("td", "1'999.6480736879855");
-    cy.get("tbody").children().eq(1).contains("td", "1'999.5937766505376");
-    cy.get("tbody").children().eq(2).contains("td", "1'999.5394796130897");
+    cy.get("tbody").children().eq(0).contains("td", "1'999.7876528649533");
+    cy.get("tbody").children().eq(1).contains("td", "1'999.3616966527707");
+    cy.get("tbody").children().eq(2).contains("td", "1'999.212604015699");
   });
 });