+++ src/client/cypress/support/e2e.js
@@ -4,3 +4,11 @@ beforeEach(() => {
   interceptApiCalls();
   loginAndResetState();
 });
+
+Cypress.on("uncaught:exception", (err, runnable) => {
+  if (
+    err.message.includes("Cannot read properties of undefined (reading 'NaN')")
+  ) {
+    return false;
+  }
+});